@@ -1,106 +1,89 @@
 {
-    "name": "shapez.io",
-    "version": "1.0.0",
+    "name": "shapez",
+    "version": "1.6.0",
     "main": "index.js",
     "repository": "https://github.com/tobspr-games/shapez.io",
     "author": "tobspr Games <hello@tobspr.io>",
-    "license": "MIT",
+    "license": "GPL-3.0-or-later",
     "private": true,
+    "type": "module",
     "scripts": {
-        "dev": "cd gulp && yarn gulp",
-        "devStandalone": "cd gulp && yarn gulp serve.standalone-steam",
-        "tslint": "cd src/js && tsc",
+        "gulp": "gulp --cwd gulp",
         "lint": "eslint src/js",
         "prettier-all": "prettier --write src/**/*.* && prettier --write gulp/**/*.*",
-        "publishOnItchWindows": "butler push tmp_standalone_files/shapez.io-standalone-win32-x64 tobspr/shapezio:windows --userversion-file version",
-        "publishOnItchLinux": "butler push tmp_standalone_files/shapez.io-standalone-linux-x64 tobspr/shapezio:linux --userversion-file version",
-        "publishOnItch": "yarn publishOnItchWindows && yarn publishOnItchLinux",
-        "publishOnSteam": "cd gulp/steampipe && ./upload.bat",
-        "publishStandalone": "yarn publishOnItch && yarn publishOnSteam",
-        "publishWeb": "cd gulp && yarn main.deploy.prod",
-        "publish": "yarn publishStandalone && yarn publishWeb",
-        "syncTranslations": "node sync-translations.js",
         "buildTypes": "tsc src/js/application.js --declaration --allowJs --emitDeclarationOnly --skipLibCheck --out types.js"
     },
     "dependencies": {
-        "@babel/core": "^7.5.4",
-        "@babel/plugin-transform-block-scoping": "^7.4.4",
-        "@babel/plugin-transform-classes": "^7.5.5",
-        "@babel/preset-env": "^7.5.4",
-        "@nastyox/rando.js": "^2.0.5",
-        "@types/cordova": "^0.0.34",
-        "@types/filesystem": "^0.0.29",
         "ajv": "^6.10.2",
-        "babel-core": "^6.26.3",
-        "babel-loader": "^8.0.4",
-        "circular-dependency-plugin": "^5.0.2",
+        "awesome-typescript-loader": "^5.2.1",
         "circular-json": "^0.5.9",
         "clipboard-copy": "^3.1.0",
-        "colors": "^1.3.3",
-        "core-js": "3",
         "crc": "^3.8.0",
-        "cssnano-preset-advanced": "^4.0.7",
         "debounce-promise": "^3.1.2",
-        "email-validator": "^2.0.4",
-        "eslint": "7.1.0",
-        "fastdom": "^1.0.8",
-        "flatted": "^2.0.1",
         "howler": "^2.1.2",
-        "html-loader": "^0.5.5",
-        "ignore-loader": "^0.1.2",
-        "logrocket": "^1.0.7",
         "lz-string": "^1.4.4",
-        "markdown-loader": "^4.0.0",
-        "match-all": "^1.2.5",
-        "phonegap-plugin-mobile-accessibility": "^1.0.5",
-        "postcss": ">=5.0.0",
-        "promise-polyfill": "^8.1.0",
-        "query-string": "^6.8.1",
         "rusha": "^0.8.13",
         "semver": "^7.3.5",
-        "strictdom": "^1.0.1",
-        "string-replace-webpack-plugin": "^0.1.3",
-        "terser-webpack-plugin": "^1.1.0",
-        "typescript": "3.9.3",
-        "uglify-template-string-loader": "^1.1.0",
-        "unused-files-webpack-plugin": "^3.4.0",
-        "webpack": "^4.43.0",
-        "webpack-bundle-analyzer": "^3.0.3",
-        "webpack-cli": "^3.1.0",
-        "webpack-deep-scope-plugin": "^1.6.0",
-        "webpack-plugin-replace": "^1.1.1",
-        "webpack-strip-block": "^0.2.0",
-        "whatwg-fetch": "^3.0.0",
-        "worker-loader": "^2.0.0",
-        "yaml": "^1.10.0",
-        "yawn-yaml": "^1.5.0"
+        "ts-loader": "^9.4.2",
+        "typescript": "^4.9.4"
     },
     "devDependencies": {
-        "@octokit/rest": "^18.0.6",
-        "@typescript-eslint/eslint-plugin": "3.0.1",
-        "@typescript-eslint/parser": "3.0.1",
-        "autoprefixer": "^9.4.3",
-        "babel-plugin-closure-elimination": "^1.3.0",
-        "babel-plugin-console-source": "^2.0.2",
-        "babel-plugin-danger-remove-unused-import": "^1.1.2",
+        "@types/circular-dependency-plugin": "^5.0.5",
+        "@types/filesystem": "^0.0.29",
+        "@types/gulp": "^4.0.9",
+        "@types/gulp-htmlmin": "^1.3.32",
+        "@types/lz-string": "^1.3.34",
+        "@types/node": "^18.11.18",
+        "@types/webpack": "^5.28.0",
+        "@typescript-eslint/eslint-plugin": "^5.42.1",
+        "@typescript-eslint/parser": "^5.42.1",
+        "browser-sync": "^2.27.10",
+        "circular-dependency-plugin": "^5.2.2",
         "css-mqpacker": "^7.0.0",
         "cssnano": "^4.1.10",
-        "eslint-config-prettier": "6.11.0",
-        "eslint-plugin-prettier": "3.1.3",
-        "faster.js": "^1.1.0",
+        "delete-empty": "^3.0.0",
+        "electron-notarize": "^1.2.1",
+        "electron-packager": "^15.4.0",
+        "eslint": "^8.27.0",
+        "eslint-config-prettier": "^8.5.0",
+        "eslint-plugin-prettier": "^4.2.1",
         "glob": "^7.1.3",
+        "gulp": "^4.0.2",
+        "gulp-audiosprite": "^1.1.0",
+        "gulp-cache": "^1.1.3",
+        "gulp-cached": "^1.1.1",
+        "gulp-clean": "^0.4.0",
+        "gulp-dart-sass": "^1.0.2",
+        "gulp-dom": "^1.0.0",
+        "gulp-fluent-ffmpeg": "^2.0.0",
+        "gulp-html-beautify": "^1.0.1",
+        "gulp-htmlmin": "^5.0.1",
+        "gulp-if": "^3.0.0",
+        "gulp-imagemin": "^7.1.0",
+        "gulp-plumber": "^1.2.1",
+        "gulp-postcss": "^8.0.0",
+        "gulp-rename": "^2.0.0",
+        "gulp-sass-lint": "^1.4.0",
+        "gulp-sftp": "git+https://git@github.com/webksde/gulp-sftp",
+        "gulp-webserver": "^0.9.1",
+        "gulp-yaml": "^2.0.4",
+        "imagemin-gifsicle": "^7.0.0",
+        "imagemin-jpegtran": "^7.0.0",
         "imagemin-mozjpeg": "^8.0.0",
         "imagemin-pngquant": "^8.0.0",
-        "jimp": "^0.6.1",
-        "js-yaml": "^3.13.1",
         "postcss-assets": "^5.0.0",
+        "postcss-critical-split": "^2.5.3",
         "postcss-preset-env": "^6.5.0",
         "postcss-round-subpixels": "^1.2.0",
-        "postcss-unprefix": "^2.1.3",
         "prettier": "^2.0.4",
-        "sass-unused": "^0.3.0",
         "strip-json-comments": "^3.0.1",
-        "trim": "^0.0.1",
+        "terser-webpack-plugin": "^5.3.6",
+        "webpack": "^5.75.0",
+        "webpack-deadcode-plugin": "^0.1.17",
+        "webpack-stream": "^7.0.0",
+        "webpack-strip-block": "^0.2.0",
+        "worker-loader": "^3.0.8",
+        "yaml": "^1.10.0",
         "yarn": "^1.22.4"
     }
 }
