@@ -2,6 +2,8 @@
 import { Application } from "../application";
 /* typehints:end */
 
+import debug from "./config.local";
+
 export const IS_DEBUG =
     G_IS_DEV &&
     typeof window !== "undefined" &&
@@ -41,8 +43,7 @@ export const THIRDPARTY_URLS = {
  */
 export function openStandaloneLink(app, campaign) {
     const discount = globalConfig.currentDiscount > 0 ? "_discount" + globalConfig.currentDiscount : "";
-    const steamSuffix = G_IS_STEAM_DEMO ? "_steamdemo" : "";
-    const event = campaign + discount + steamSuffix;
+    const event = campaign + discount;
     app.platformWrapper.openExternalLink(THIRDPARTY_URLS.standaloneCampaignLink.replace("$campaign", event));
     app.gameAnalytics.noteMinor("g.stdlink." + event);
 }
@@ -129,7 +130,7 @@ export const globalConfig = {
     },
 
     rendering: {},
-    debug: require("./config.local").default,
+    debug,
 
     currentDiscount: 0,
 
