@@ -58,9 +58,10 @@ export class StaleAreaDetector {
             for (let i = 0; i < componentIds.length; ++i) {
                 if (entity.components[componentIds[i]]) {
                     // Entity is relevant, compute affected area
-                    const area = entity.components.StaticMapEntity.getTileSpaceBounds().expandedInAllDirections(
-                        tilesAround
-                    );
+                    const area =
+                        entity.components.StaticMapEntity.getTileSpaceBounds().expandedInAllDirections(
+                            tilesAround
+                        );
                     this.invalidate(area);
                     return;
                 }
