@@ -1,5 +1,4 @@
 import { T } from "../translations";
-import { rando } from "@nastyox/rando.js";
 import { WEB_STEAM_SSO_AUTHENTICATED } from "./steam_sso";
 
 const bigNumberSuffixTranslationKeys = ["thousands", "millions", "billions", "trillions"];
@@ -44,7 +43,7 @@ export function newEmptyMap() {
  * @param {number} end
  */
 export function randomInt(start, end) {
-    return rando(start, end);
+    return Math.floor(Math.random() * (end - start + 1) + start);
 }
 
 /**
@@ -694,10 +693,6 @@ const romanLiteralsCache = ["0"];
  * @returns {string}
  */
 export function getRomanNumber(number) {
-    if (G_WEGAME_VERSION) {
-        return String(number);
-    }
-
     number = Math.max(0, Math.round(number));
     if (romanLiteralsCache[number]) {
         return romanLiteralsCache[number];
@@ -753,18 +748,6 @@ export function getRomanNumber(number) {
  * Returns the appropriate logo sprite path
  */
 export function getLogoSprite() {
-    if (G_WEGAME_VERSION) {
-        return "logo_wegame.png";
-    }
-
-    if (G_IS_STEAM_DEMO) {
-        return "logo_demo.png";
-    }
-
-    if (G_CHINA_VERSION) {
-        return "logo_cn.png";
-    }
-
     if (G_IS_STANDALONE || WEB_STEAM_SSO_AUTHENTICATED) {
         return "logo.png";
     }
