@@ -123,18 +123,16 @@ export class HUDShop extends BaseHUDPart {
                 container.appendChild(pinButton);
 
                 let infoDetector;
-                if (!G_WEGAME_VERSION) {
-                    const viewInfoButton = document.createElement("button");
-                    viewInfoButton.classList.add("showInfo");
-                    container.appendChild(viewInfoButton);
-                    infoDetector = new ClickDetector(viewInfoButton, {
-                        consumeEvents: true,
-                        preventDefault: true,
-                    });
-                    infoDetector.click.add(() =>
-                        this.root.hud.signals.viewShapeDetailsRequested.dispatch(shapeDef)
-                    );
-                }
+                const viewInfoButton = document.createElement("button");
+                viewInfoButton.classList.add("showInfo");
+                container.appendChild(viewInfoButton);
+                infoDetector = new ClickDetector(viewInfoButton, {
+                    consumeEvents: true,
+                    preventDefault: true,
+                });
+                infoDetector.click.add(() =>
+                    this.root.hud.signals.viewShapeDetailsRequested.dispatch(shapeDef)
+                );
 
                 const currentGoalShape = this.root.hubGoals.currentGoal.definition.getHash();
                 if (shape === currentGoalShape) {
