@@ -99,10 +99,12 @@ export class ShapeDefinitionManager extends BasicSerializableObject {
 
         this.root.signals.achievementCheck.dispatch(ACHIEVEMENTS.cutShape, null);
 
-        return /** @type {[ShapeDefinition, ShapeDefinition]} */ (this.operationCache[key] = [
-            this.registerOrReturnHandle(rightSide),
-            this.registerOrReturnHandle(leftSide),
-        ]);
+        return /** @type {[ShapeDefinition, ShapeDefinition]} */ (
+            this.operationCache[key] = [
+                this.registerOrReturnHandle(rightSide),
+                this.registerOrReturnHandle(leftSide),
+            ]
+        );
     }
 
     /**
@@ -113,18 +115,19 @@ export class ShapeDefinitionManager extends BasicSerializableObject {
     shapeActionCutQuad(definition) {
         const key = "cut-quad/" + definition.getHash();
         if (this.operationCache[key]) {
-            return /** @type {[ShapeDefinition, ShapeDefinition, ShapeDefinition, ShapeDefinition]} */ (this
-                .operationCache[key]);
+            return /** @type {[ShapeDefinition, ShapeDefinition, ShapeDefinition, ShapeDefinition]} */ (
+                this.operationCache[key]
+            );
         }
 
-        return /** @type {[ShapeDefinition, ShapeDefinition, ShapeDefinition, ShapeDefinition]} */ (this.operationCache[
-            key
-        ] = [
-            this.registerOrReturnHandle(definition.cloneFilteredByQuadrants([0])),
-            this.registerOrReturnHandle(definition.cloneFilteredByQuadrants([1])),
-            this.registerOrReturnHandle(definition.cloneFilteredByQuadrants([2])),
-            this.registerOrReturnHandle(definition.cloneFilteredByQuadrants([3])),
-        ]);
+        return /** @type {[ShapeDefinition, ShapeDefinition, ShapeDefinition, ShapeDefinition]} */ (
+            this.operationCache[key] = [
+                this.registerOrReturnHandle(definition.cloneFilteredByQuadrants([0])),
+                this.registerOrReturnHandle(definition.cloneFilteredByQuadrants([1])),
+                this.registerOrReturnHandle(definition.cloneFilteredByQuadrants([2])),
+                this.registerOrReturnHandle(definition.cloneFilteredByQuadrants([3])),
+            ]
+        );
     }
 
     /**
@@ -142,9 +145,9 @@ export class ShapeDefinitionManager extends BasicSerializableObject {
 
         this.root.signals.achievementCheck.dispatch(ACHIEVEMENTS.rotateShape, null);
 
-        return /** @type {ShapeDefinition} */ (this.operationCache[key] = this.registerOrReturnHandle(
-            rotated
-        ));
+        return /** @type {ShapeDefinition} */ (
+            this.operationCache[key] = this.registerOrReturnHandle(rotated)
+        );
     }
 
     /**
@@ -160,9 +163,9 @@ export class ShapeDefinitionManager extends BasicSerializableObject {
 
         const rotated = definition.cloneRotateCCW();
 
-        return /** @type {ShapeDefinition} */ (this.operationCache[key] = this.registerOrReturnHandle(
-            rotated
-        ));
+        return /** @type {ShapeDefinition} */ (
+            this.operationCache[key] = this.registerOrReturnHandle(rotated)
+        );
     }
 
     /**
@@ -178,9 +181,9 @@ export class ShapeDefinitionManager extends BasicSerializableObject {
 
         const rotated = definition.cloneRotate180();
 
-        return /** @type {ShapeDefinition} */ (this.operationCache[key] = this.registerOrReturnHandle(
-            rotated
-        ));
+        return /** @type {ShapeDefinition} */ (
+            this.operationCache[key] = this.registerOrReturnHandle(rotated)
+        );
     }
 
     /**
@@ -198,9 +201,9 @@ export class ShapeDefinitionManager extends BasicSerializableObject {
         this.root.signals.achievementCheck.dispatch(ACHIEVEMENTS.stackShape, null);
 
         const stacked = lowerDefinition.cloneAndStackWith(upperDefinition);
-        return /** @type {ShapeDefinition} */ (this.operationCache[key] = this.registerOrReturnHandle(
-            stacked
-        ));
+        return /** @type {ShapeDefinition} */ (
+            this.operationCache[key] = this.registerOrReturnHandle(stacked)
+        );
     }
 
     /**
@@ -218,9 +221,9 @@ export class ShapeDefinitionManager extends BasicSerializableObject {
         this.root.signals.achievementCheck.dispatch(ACHIEVEMENTS.paintShape, null);
 
         const colorized = definition.cloneAndPaintWith(color);
-        return /** @type {ShapeDefinition} */ (this.operationCache[key] = this.registerOrReturnHandle(
-            colorized
-        ));
+        return /** @type {ShapeDefinition} */ (
+            this.operationCache[key] = this.registerOrReturnHandle(colorized)
+        );
     }
 
     /**
@@ -235,9 +238,9 @@ export class ShapeDefinitionManager extends BasicSerializableObject {
             return /** @type {ShapeDefinition} */ (this.operationCache[key]);
         }
         const colorized = definition.cloneAndPaintWith4Colors(colors);
-        return /** @type {ShapeDefinition} */ (this.operationCache[key] = this.registerOrReturnHandle(
-            colorized
-        ));
+        return /** @type {ShapeDefinition} */ (
+            this.operationCache[key] = this.registerOrReturnHandle(colorized)
+        );
     }
 
     /**
@@ -261,9 +264,9 @@ export class ShapeDefinitionManager extends BasicSerializableObject {
      * @returns {ShapeDefinition}
      */
     getDefinitionFromSimpleShapes(subShapes, color = enumColors.uncolored) {
-        const shapeLayer = /** @type {import("./shape_definition").ShapeLayer} */ (subShapes.map(
-            subShape => ({ subShape, color })
-        ));
+        const shapeLayer = /** @type {import("./shape_definition").ShapeLayer} */ (
+            subShapes.map(subShape => ({ subShape, color }))
+        );
 
         return this.registerOrReturnHandle(new ShapeDefinition({ layers: [shapeLayer] }));
     }
