@@ -28,6 +28,12 @@ export const MOD_SIGNALS = {
 
     stateEntered: /** @type {TypedSignal<[GameState]>} */ (new Signal()),
 
-    gameSerialized: /** @type {TypedSignal<[GameRoot, import("../savegame/savegame_typedefs").SerializedGame]>} */ (new Signal()),
-    gameDeserialized: /** @type {TypedSignal<[GameRoot, import("../savegame/savegame_typedefs").SerializedGame]>} */ (new Signal()),
+    gameSerialized:
+        /** @type {TypedSignal<[GameRoot, import("../savegame/savegame_typedefs").SerializedGame]>} */ (
+            new Signal()
+        ),
+    gameDeserialized:
+        /** @type {TypedSignal<[GameRoot, import("../savegame/savegame_typedefs").SerializedGame]>} */ (
+            new Signal()
+        ),
 };
