@@ -30,29 +30,16 @@ export class SettingsState extends TextualGameState {
                     : ""
             }
 
-            ${
-                G_WEGAME_VERSION
-                    ? ""
-                    : `
-                <button class="styledButton categoryButton manageMods">${T.mods.title}
-                    <span class="newBadge">${T.settings.newBadge}</span>
-                </button>`
-            }
+            <button class="styledButton categoryButton manageMods">${T.mods.title}
+                <span class="newBadge">${T.settings.newBadge}</span>
+            </button>
 
 
-            <div class="other ${G_CHINA_VERSION || G_WEGAME_VERSION ? "noabout" : ""}">
-
-            ${
-                G_CHINA_VERSION || G_WEGAME_VERSION
-                    ? ""
-                    : `
+            <div class="other">
                 <button class="styledButton about">${T.about.title}</button>
                 <button class="styledButton privacy">Privacy Policy</button>
-
-`
-            }
                 <div class="versionbar">
-                    ${G_WEGAME_VERSION ? "" : `<div class="buildVersion">${T.global.loading} ...</div>`}
+                    <div class="buildVersion">${T.global.loading} ...</div>
                 </div>
             </div>
         </div>
@@ -122,14 +109,12 @@ export class SettingsState extends TextualGameState {
     onEnter(payload) {
         this.renderBuildText();
 
-        if (!G_CHINA_VERSION && !G_WEGAME_VERSION) {
-            this.trackClicks(this.htmlElement.querySelector(".about"), this.onAboutClicked, {
-                preventDefault: false,
-            });
-            this.trackClicks(this.htmlElement.querySelector(".privacy"), this.onPrivacyClicked, {
-                preventDefault: false,
-            });
-        }
+        this.trackClicks(this.htmlElement.querySelector(".about"), this.onAboutClicked, {
+            preventDefault: false,
+        });
+        this.trackClicks(this.htmlElement.querySelector(".privacy"), this.onPrivacyClicked, {
+            preventDefault: false,
+        });
 
         const keybindingsButton = this.htmlElement.querySelector(".editKeybindings");
 
